airflow/requirements.txt
xmltodict
kafka-python
pymongo
apache-airflow-providers-apache-spark
sqlalchemy